requirements.txt
absl-py==0.6.1
astor==0.7.1
certifi==2019.3.9
chardet==3.0.4
cloudpickle==0.6.1
colorama==0.4.1
dask==1.0.0
decorator==4.3.0
gast==0.2.0
grpcio==1.16.1
h5py==2.8.0
idna==2.8
imageio==2.4.1
Keras==2.2.4
Keras-Applications==1.0.6
Keras-Preprocessing==1.0.5
Markdown==3.0.1
networkx==2.2
numpy==1.15.4
Pillow==5.3.0
protobuf==3.6.1
PyWavelets==1.0.1
PyYAML==3.13
qrcode==6.0
requests==2.21.0
scikit-image==0.14.1
scipy==1.1.0
six==1.11.0
tensorboard==1.12.0
tensorflow==1.12.0
termcolor==1.1.0
toolz==0.9.0
tqdm==4.28.1
urllib3==1.24.1
Werkzeug==0.14.1
